+++ historic/demos/src/rooms/blockfall/blockfall.js
@@ -43,7 +43,7 @@ export const WrappingWorld = SpaceWrapping({
 })(PhysicalWorld);
 
 export default function initBlockfall(options) {
-    // called as part of installing the initial Island
+    // called as part of installing the initial VirtualMachine
     const room = Room.create();
     room.addElementManipulators = false;
     room.noNavigation = true;