+++ packages/croquet/teatime/index.js
@@ -8,7 +8,7 @@ export { App } from "@croquet/util/html";
 // unofficial exports
 export { default as Controller } from "./src/controller";
 export { currentRealm } from "./src/realms";
-export { QFunc, gatherInternalClassTypes } from "./src/island";
+export { QFunc, gatherInternalClassTypes } from "./src/vm";
 export { Messenger }  from "./src/messenger";
 
 // putting event documentation here because JSDoc errors when parsing controller.js at the moment