+++ packages/croquet/CHANGE_LOG.md
@@ -241,4 +241,4 @@ If someone feels like going through these changes, please do so.
 * 2019-03-06: + Reflector
 * 2019-03-01: + Future Messages
 * 2019-02-22: + Serialization
-* 2019-02-16: Model + View + Island
+* 2019-02-16: Model + View + VirtualMachine